CreatureTest corrected.
MovementplannerTest commented because it caused more than 100 errors. and it wouldnt compile.

Creature:
Changed weightChild in Weight
Added steps taken to match the speed and made it work when the last step was bigger than the steps in the list.

Removed "findCreature" and "findPlant" from Movement planner.

World:
Added an option for swimming, didn't implement it jet.